Handle framerate

+++ client/src/main/scala/domain/components/Position.scala
@@ -1,3 +1,3 @@
 package domain.components
 
-case class Position(x: Int, y: Int)
+case class Position(x: Double, y: Double)
