+++ client/src/main/scala/utils/Utility.scala
@@ -1,7 +1,7 @@
 package utils
 
 object Utility {
-  def positiveModulo(a: Int, n: Int): Int = {
+  def positiveModulo(a: Double, n: Double): Double = {
     // to get positive mod
     // we compute remainder, and add back the n
     // to prevent result bigger than n which violate the rules
